+++ background.js
@@ -69,6 +69,8 @@ const configs = {
     'isMac': false,
     'enableContainerIdentify': true,
 
+    'showContextMenuItem': false,
+
     'linkHint': false,
     'linkDisabledUrls': [],
 
@@ -92,6 +94,7 @@ Wikipedia=>https://wikipedia.org/w/index.php?title=Special:Search&search=%s`,
 
 let openPopups = [];
 let activePopupCount = 0;
+let lastContextX, lastContextY;
 
 // Load user configurations from storage
 async function loadUserConfigs() {
@@ -203,19 +206,48 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
                 chrome.contextMenus.onClicked.removeListener(onMenuItemClicked);
                 chrome.contextMenus.onClicked.addListener(onMenuItemClicked);
 
+                loadUserConfigs().then(userConfigs => {
+                    if (userConfigs.showContextMenuItem) {
+                        lastContextX = request.lastClientX;
+                        lastContextY = request.lastClientY;
+                        chrome.contextMenus.remove('showContextMenuItem', () => {
+                            if (chrome.runtime.lastError) {
+                                // console.error("Error removing context menu: ", chrome.runtime.lastError.message);
+                            }
+                        });
+                        chrome.contextMenus.create({
+                            id: 'showContextMenuItem',
+                            title: chrome.i18n.getMessage('previewItem'),
+                            contexts: ['link']
+                        });
+                    }
+                    if (userConfigs.showContextMenuItem === false) {
+                        chrome.contextMenus.remove('showContextMenuItem', () => {
+                            if (chrome.runtime.lastError) {
+                                // console.error("Error removing context menu: ", chrome.runtime.lastError.message);
+                            }
+                        });
+                    }
+                });
+
                 if (typeof request.addContextMenuItem !== "undefined") {
-                    chrome.contextMenus.remove('sendPageBack', () => {
-                        if (chrome.runtime.lastError) {
-                            // console.error("Error removing context menu: ", chrome.runtime.lastError.message);
-                        } else {
-                            // console.log("Context menu 'sendPageBack' removed successfully.");
-                        }
-                    });
+
+                    if (!request.addContextMenuItem) {
+                        chrome.contextMenus.remove('sendPageBack', () => {
+                            console.log('Removing context menu item if exists.');
+                            if (chrome.runtime.lastError) {
+                                // console.error("Error removing context menu: ", chrome.runtime.lastError.message);
+                            } else {
+                                // console.log("Context menu 'sendPageBack' removed successfully.");
+                            }
+                        });
+                    }
                     if (request.addContextMenuItem) {
                         chrome.storage.local.get('popupWindowsInfo', (result) => {
                             const popupWindowsInfo = result.popupWindowsInfo || {};
 
                             loadUserConfigs().then(userConfigs => {
+                                console.log('Creating context menu item for sending page back.');
                                 chrome.contextMenus.create({
                                     id: 'sendPageBack',
                                     title: chrome.i18n.getMessage('sendPageBack'),
@@ -1031,6 +1063,12 @@ function onMenuItemClicked(info, tab) {
                 console.error('popupWindowsInfo is empty or not properly structured.');
             }
         });
+    }
+
+    if (info.menuItemId === 'showContextMenuItem') {
+
+        // Send message to content script to handle preview
+        chrome.tabs.sendMessage(tab.id, { linkUrl: info.linkUrl, trigger: 'contextMenu', x: lastContextX, y: lastContextY });
     }
 }